Create a simple tutorial for neural machine transl
Create a simple tutorial for neural machine translation. This was done a bit quickly, so we should check it against a larger data set and refine names/language in order to make it as simple as possible. We should also illustrate how more than one translation can come from such a model (with one of the strategies we have discussed).